refactor: Rename `Client` to `RedisClient`
